temp

+++ decompiler/util/data_decompile.cpp
@@ -1043,7 +1043,7 @@ goos::Object decompile_pair_elt(const LinkedWord& word,
   } else if (word.kind == LinkedWord::EMPTY_PTR) {
     return pretty_print::to_symbol("'()");
   } else if (word.kind == LinkedWord::PLAIN_DATA && (word.data & 0b111) == 0) {
-    return pretty_print::to_symbol(fmt::format("(the binteger {})", word.data >> 3));
+    return pretty_print::to_symbol(fmt::format("(the binteger {})", ((s32)word.data) >> 3));
   } else {
     throw std::runtime_error(fmt::format("Pair elt did not have a good word kind"));
   }
